feat: upgrade to Rocky 10
Upgrade the image build to use Rocky Linux 10.
Remove br_netfilter Kernel module related tasks because everything uses
nftables these days and the task failed on Rocky 10.
Pin Calico version being installed in the test and install it in it's
own namespace because of a RBAC issue in the Tigera Operator chart.
Refactor the test to validate Pod readiness in all namespaces.